+++ CHANGELOG.md
@@ -26,12 +26,17 @@
 - Added functions to `MastForest` and `MastForestBuilder` to add and ensure nodes with fewer LOC (#1404, #1412)
 - Added `Assembler::assemble_library()` (#1413).
 - Added `Assembler::assemble_kernel()` (#1418).
+<<<<<<< HEAD
 - Added `miden_core::prettier::pretty_print_csv` helper, for formatting of iterators over `PrettyPrint` values as comma-separated items
+=======
+- Added source location tracking to assembled MAST (#1419)
+>>>>>>> 32381c6a (feat(assembly): track source locations in debug mode)
 
 #### Changed
 
 - When using `if.(true|false) .. end`, the parser used to emit an empty block for the branch that was elided. The parser now emits a block containing a single `nop` instruction instead, which is equivalent to the code emitted by the assembler when lowering to MAST.
 - `internals` configuration feature was renamed to `testing` (#1399).
+- The `AssemblyOp` decorator now contains an optional `SourceLocation` (#1419)
 
 ## 0.9.2 (2024-05-22) - `stdlib` crate only